reorg: folder structure cleanup (all steps from #1968)
