+++ Sources/RunBotCore/Platform/LogFetcher.swift
