+++ Sources/RunBotCore/WorkflowScope/ScopeEntry.swift
