+++ Sources/RunBotCore/WorkflowScope/ScopePreferences.swift
