+++ Sources/RunBotCore/WorkflowScope/ScopePreferencesStore.swift
