+++ Sources/RunBotCore/WorkflowScope/ScopeStoreProtocol.swift
